refactor(parts): remove duplicate test DTOs for ApiPart and PartRequest
Add a parameterless constructor to the production SheetMusic.Api.Parts.ViewModels.ApiPart
view model so it can be deserialized directly by System.Text.Json in
integration tests, alongside the existing MusicPart-based constructor
used by production code (behavior unchanged for existing call sites).

Test code (PartTests, SetTests, PartDataBuilder, FakerFactory) now
references the production ApiPart and PartRequest types directly
instead of maintaining separate test-only duplicates.
