--- src/SheetMusic.Api.Test/Parts/Models/ApiPart.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace SheetMusic.Api.Test.Parts.Models;
-
-public class ApiPart
-{
-    public Guid Id { get; set; }
-
-    public string Name { get; set; } = null!;
-
-    public int SortOrder { get; set; }
-
-    public bool Indexable { get; set; }
-
-    public List<string> Aliases { get; set; } = new List<string>();
-}
