--- src/SheetMusic.Api.Test/Parts/Models/PutPartModel.cs
@@ -1,10 +1,0 @@
-﻿namespace SheetMusic.Api.Test.Parts.Models;
-
-public class PutPartModel
-{
-    public string Name { get; set; } = null!;
-
-    public int SortOrder { get; set; }
-
-    public bool? Indexable { get; set; }
-}
