+++ src/SheetMusic.Api.Test/Tests/SetTests.cs
@@ -1,11 +1,11 @@
 ﻿using FluentAssertions;
 using Moq;
 using SheetMusic.Api.BlobStorage;
+using SheetMusic.Api.Parts.ViewModels;
 using SheetMusic.Api.Test.Infrastructure;
 using SheetMusic.Api.Test.Infrastructure.Authentication;
 using SheetMusic.Api.Test.Infrastructure.TestCollections;
 using SheetMusic.Api.Test.Models;
-using SheetMusic.Api.Test.Parts.Models;
 using SheetMusic.Api.Test.Utility;
 using System.Collections.Generic;
 using System.Diagnostics;
